+++ src/test/java/kitchenpos/menu/acceptance/step/MenuAcceptanceStep.java
@@ -1,4 +1,4 @@
-package kitchenpos.acceptance.step;
+package kitchenpos.menu.acceptance.step;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertAll;
@@ -8,6 +8,7 @@ import io.restassured.response.ExtractableResponse;
 import io.restassured.response.Response;
 import java.math.BigDecimal;
 import java.util.List;
+import kitchenpos.acceptance.step.HttpUtil;
 import kitchenpos.menu.dto.MenuRequest;
 import kitchenpos.menu.dto.MenuResponse;
